Feature/us012 add unit test to services controllers

+++ scr/Parking-Zone.MVC/Areas/Admin/Controllers/ParkingZonesController.cs
@@ -1,8 +1,6 @@
-﻿using AutoMapper;
-using Microsoft.AspNetCore.Authorization;
+﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
-using Parking_Zone.Domain.Entities;
 using Parking_Zone.MVC.ViewModels.ParkingZoneVMs;
 using Parking_Zone.Service.Interfaces;
 namespace Parking_Zone.MVC.Areas.Admin.Controllers;
@@ -12,37 +10,33 @@ namespace Parking_Zone.MVC.Areas.Admin.Controllers;
 public class ParkingZonesController : Controller
 {
     private readonly IParkingZoneService _parkingZoneService;
-    private readonly IMapper _mapper;
+    private readonly IAddressService _addressService;
 
-    public ParkingZonesController(IParkingZoneService parkingZoneService, IMapper mapper)
+    public ParkingZonesController(IParkingZoneService parkingZoneService, IAddressService addressService)
     {
         _parkingZoneService = parkingZoneService;
-        _mapper = mapper;
+        _addressService = addressService;
     }
 
     // GET: Admin/ParkingZones
     public IActionResult Index()
     {
         var parkingZones = _parkingZoneService.RetrieveAll();
-        ParkingZoneIndexVM resVM = new()
-        { ParkingZones = parkingZones };
-        return View(resVM);
+        var VMs = parkingZones.Select(x=> new ListOfItemsVM(x));
+        return View(VMs);
     }
 
     // GET: Admin/ParkingZones/Details/5
     public IActionResult Details(long? id)
     {
-        if (id == null)
-            return NotFound();
-
         var parkingZone = _parkingZoneService.RetrieveById(id);
 
-        if (parkingZone == null)
+        if (parkingZone is null)
             return NotFound();
 
-        ParkingZoneDetailsVM resVM = new();
+        DetailsVM resVM = new DetailsVM(parkingZone);
 
-        return View(_mapper.Map(parkingZone,resVM));
+        return View(resVM);
     }
 
     // GET: Admin/ParkingZones/Create
@@ -54,12 +48,11 @@ public class ParkingZonesController : Controller
     // POST: Admin/ParkingZones/Create
     [HttpPost]
     [ValidateAntiForgeryToken]
-    public IActionResult Create(ParkingZoneCreateVM createVM)
+    public IActionResult Create(CreateVM createVM)
     {
         if (ModelState.IsValid)
         {
-            var parkingZone = new ParkingZone();
-            _parkingZoneService.Insert(_mapper.Map(createVM,parkingZone));
+            _parkingZoneService.Insert(createVM.MapToModel());
 
             return RedirectToAction(nameof(Index));
         }
@@ -69,33 +62,31 @@ public class ParkingZonesController : Controller
     // GET: Admin/ParkingZones/Edit/5
     public IActionResult Edit(long? id)
     {
-        if (id == null)
-            return NotFound();
-
         var parkingZone = _parkingZoneService.RetrieveById(id);
 
-        if (parkingZone == null)
+        if (parkingZone is null)
             return NotFound();
 
-        ParkingZoneEditVM resVM = new();
+        EditVM resVM = new(parkingZone);
 
-        return View(_mapper.Map(parkingZone,resVM));
+        return View(resVM);
     }
 
     // POST: Admin/ParkingZones/Edit/5
     [HttpPost]
     [ValidateAntiForgeryToken]
-    public IActionResult Edit(long id, ParkingZoneEditVM parkingZoneEditVM)
+    public IActionResult Edit(long id, EditVM parkingZoneEditVM)
     {
-        if (id != parkingZoneEditVM.Id)
+        var existingParkingZone = _parkingZoneService.RetrieveById(id);
+
+        if (id != existingParkingZone.Id)
             return NotFound();
 
         if (ModelState.IsValid)
         {
             try
             {
-                var parkingZone = _parkingZoneService.RetrieveById(id);
-                _parkingZoneService.Modify(id, _mapper.Map(parkingZoneEditVM,parkingZone));
+                _parkingZoneService.Modify(parkingZoneEditVM.MapToModel(existingParkingZone));
             }
             catch (DbUpdateConcurrencyException)
             {
@@ -112,16 +103,11 @@ public class ParkingZonesController : Controller
     // GET: Admin/ParkingZones/Delete/5
     public IActionResult Delete(long? id)
     {
-        if (id == null)
-            return NotFound();
-
         var parkingZone = _parkingZoneService.RetrieveById(id);
-        if (parkingZone == null)
+        if (parkingZone is null)
             return NotFound();
 
-        var resVM = new ParkingZoneDeleteVM();
-
-        return View(_mapper.Map(parkingZone,resVM));
+        return View(parkingZone);
     }
 
     // POST: Admin/ParkingZones/Delete/5
@@ -129,7 +115,12 @@ public class ParkingZonesController : Controller
     [ValidateAntiForgeryToken]
     public IActionResult DeleteConfirmed(long id)
     {
+        var existingParkingZone = _parkingZoneService.RetrieveById(id);
+        if (existingParkingZone is null)
+            return NotFound();
+
         _parkingZoneService.Remove(id);
+        _addressService.Remove(existingParkingZone.AddressId);
 
         return RedirectToAction(nameof(Index));
     }
@@ -138,7 +129,7 @@ public class ParkingZonesController : Controller
     {
         var park = _parkingZoneService.RetrieveById(id);
 
-        if (park == null)
+        if (park is null)
             return false;
 
         return true;
